Add files via upload
Chat-UP!
Chat-UP! is a secure chat application using Python and the TKinter library. It uses an SQLite database to store data and socket programming to transmit messages over a network. User authentication is handled using OpenLDAP and socket layer security is provided through the use of x509 certificates.

Getting Started
To use Chat-UP!, follow the instructions below.

Prerequisites
Python 3.x installed on your system.
Running the Application
Start the server by running the server.py file.
Run the main_interface.py file to open the main interface of Chat-UP!.
In the main interface, you can register a new user by clicking on the "Register" button. Fill in the required information in the registration form and submit it.
After registering, you can log in by clicking on the "Login" button. Enter your username and password in the login form and click on the "Login" button.
Once logged in, you will be able to see the list of connected users in the chat application.
Select a user from the list and click on the "Chat" button to start a conversation with that user.
In the chat window, type your message in the input box at the bottom and click on the "Send" button to send the message to the selected user.
The message will be displayed in the chat window, and you can continue the conversation by sending more messages.
Files
The Chat-UP! application consists of the following files:

main_interface.py: This file contains the code for the main interface of the chat application.
login.py: This file contains the code for the login form.
reg.py: This file contains the code for the registration form.
client.py: This file contains the code for the client-side of the chat application.
server.py: This file contains the code for the server-side of the chat application.
!Note: Make sure to run the server.py file before running the other files to ensure proper communication between the client and server.!